removed debug lines
Tested on all spells available to Quel'thalas with no anomalies reported / noticed.
Power is removed when sunwell is lost.
Tested with passives that do magic damage such as Incinerate
Tested with autocast abilities  such as Frost Arrows and the % bonus is correctly only applied to the magic damage and not the overall damage.